# openapi.yaml
openapi: 3.0.3
info:
  title: Photobook API
  description: API for managing photo albums with multi-user access control
  version: 1.0.0
  contact:
    name: Stefan Remer
    email: stefan@remer.cc

servers:
  - url: http://localhost:8081
    description: Local development server

tags:
  - name: Authentication
    description: User authentication and token management
  - name: Users
    description: User management operations
  - name: Albums
    description: Album management and access control
  - name: Photos
    description: Photo upload, management, and serving

security:
  - bearerAuth: []

paths:
  # Authentication Endpoints
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Endpoints
  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update current user profile
      operationId: updateCurrentUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users:
    get:
      tags:
        - Users
      summary: List all users (admin only)
      operationId: listUsers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          description: Forbidden - admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Users
      summary: Create new user (admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{userId}:
    delete:
      tags:
        - Users
      summary: Delete user (admin only)
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted
        '403':
          description: Forbidden - admin only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Album Endpoints
  /api/albums:
    get:
      tags:
        - Albums
      summary: List albums accessible by current user
      operationId: listAlbums
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Albums retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Albums
      summary: Create new album
      operationId: createAlbum
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumRequest'
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/albums/{albumId}:
    get:
      tags:
        - Albums
      summary: Get album details
      operationId: getAlbum
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Album retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '403':
          description: Forbidden - no access to album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Albums
      summary: Update album
      operationId: updateAlbum
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumRequest'
      responses:
        '200':
          description: Album updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not album owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Albums
      summary: Delete album
      operationId: deleteAlbum
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Album deleted
        '403':
          description: Forbidden - not album owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/albums/{albumId}/users:
    get:
      tags:
        - Albums
      summary: List users with access to album
      operationId: listAlbumUsers
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Album users retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumUserResponse'
        '403':
          description: Forbidden - not album owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Albums
      summary: Grant user access to album
      operationId: addAlbumUser
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAlbumUserRequest'
      responses:
        '201':
          description: User access granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumUserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not album owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/albums/{albumId}/users/{userId}:
    delete:
      tags:
        - Albums
      summary: Remove user access from album
      operationId: removeAlbumUser
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User access removed
        '403':
          description: Forbidden - not album owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Photo Endpoints
  /api/albums/{albumId}/photos:
    get:
      tags:
        - Photos
      summary: List photos in album
      operationId: listPhotos
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Photos retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoListResponse'
        '403':
          description: Forbidden - no access to album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Photos
      summary: Upload photos to album
      operationId: uploadPhotos
      parameters:
        - name: albumId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Photos uploaded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhotoResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - no write access to album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/photos/{photoId}:
    get:
      tags:
        - Photos
      summary: Get photo details
      operationId: getPhoto
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Photo retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
        '403':
          description: Forbidden - no access to photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Photos
      summary: Delete photo
      operationId: deletePhoto
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Photo deleted
        '403':
          description: Forbidden - no write access to album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/photos/{photoId}/url:
    get:
      tags:
        - Photos
      summary: Get photo URL (presigned)
      operationId: getPhotoUrl
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          schema:
            type: string
            enum: [small, medium, large, original]
            default: original
      responses:
        '200':
          description: Photo URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoUrlResponse'
        '403':
          description: Forbidden - no access to photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/photos/{photoId}/move:
    post:
      tags:
        - Photos
      summary: Move photo to another album
      operationId: movePhoto
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovePhotoRequest'
      responses:
        '200':
          description: Photo moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - no write access to albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo or album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/photos/{photoId}/copy:
    post:
      tags:
        - Photos
      summary: Copy photo to another album
      operationId: copyPhoto
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyPhotoRequest'
      responses:
        '201':
          description: Photo copied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - no write access to target album
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Photo or album not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Authentication Schemas
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    LogoutRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/UserResponse'

    # User Schemas
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
          default: USER

    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
          format: password

    UserListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    # Album Schemas
    AlbumResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        coverPhotoId:
          type: string
          format: uuid
          nullable: true
        photoCount:
          type: integer
        ownerId:
          type: string
          format: uuid
        ownerName:
          type: string
        role:
          type: string
          enum: [OWNER, VIEWER]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateAlbumRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string

    UpdateAlbumRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        coverPhotoId:
          type: string
          format: uuid
          nullable: true

    AlbumListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AlbumResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    AlbumUserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [OWNER, VIEWER]
        addedAt:
          type: string
          format: date-time

    AddAlbumUserRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [VIEWER]
          default: VIEWER

    # Photo Schemas
    PhotoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        albumId:
          type: string
          format: uuid
        originalFilename:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
        width:
          type: integer
        height:
          type: integer
        takenAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PROCESSING, READY, ERROR]
        metadata:
          type: object
          additionalProperties: true
        uploadedAt:
          type: string
          format: date-time

    PhotoListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PhotoResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    PhotoUrlResponse:
      type: object
      properties:
        url:
          type: string
        expiresAt:
          type: string
          format: date-time

    MovePhotoRequest:
      type: object
      required:
        - targetAlbumId
      properties:
        targetAlbumId:
          type: string
          format: uuid

    CopyPhotoRequest:
      type: object
      required:
        - targetAlbumId
      properties:
        targetAlbumId:
          type: string
          format: uuid

    # Common Schemas
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string